+++ lib/components/Tiles/Tiles.docs.tsx
@@ -42,12 +42,17 @@ const docs: ComponentDocs = {
         <Text>
           The number of tiles in each row. Accepts a number from 1 to 5, also
           supporting responsive values, e.g.{' '}
-          <Strong>{'columns={{ mobile: 2, tablet: 3, desktop: 4 }}'}</Strong>
+          <Strong>
+            {'columns={{ mobile: 2, tablet: 3, desktop: 4, wide: 5 }}'}
+          </Strong>
         </Text>
       ),
       Example: () =>
         source(
-          <Tiles space="small" columns={{ mobile: 2, tablet: 3, desktop: 4 }}>
+          <Tiles
+            space="small"
+            columns={{ mobile: 2, tablet: 3, desktop: 4, wide: 5 }}
+          >
             <Placeholder height={40} />
             <Placeholder height={40} />
             <Placeholder height={40} />
@@ -68,7 +73,9 @@ const docs: ComponentDocs = {
           Control the amount of space between each tile using the{' '}
           <Strong>space</Strong> prop, also supporting responsive values, e.g.{' '}
           <Strong>
-            {"space={{ mobile: 'small', tablet: 'medium', desktop: 'large' }}"}
+            {
+              "space={{ mobile: 'small', tablet: 'medium', desktop: 'large', wide: 'xlarge' }}"
+            }
           </Strong>
         </Text>
       ),