+++ lib/css/negativeMargin/negativeMarginTop.css.ts
@@ -28,3 +28,9 @@ export const desktop = styleVariants({ none: 0, ...vars.space }, (value) =>
     desktop: negativeMarginTop(value),
   }),
 );
+
+export const wide = styleVariants({ none: 0, ...vars.space }, (value) =>
+  responsiveStyle({
+    wide: negativeMarginTop(value),
+  }),
+);