+++ lib/utils/resolveResponsiveRangeProps.ts
@@ -1,17 +1,17 @@
 import { Breakpoint, breakpointNames } from '../css/breakpoints';
 
 export interface ResponsiveRangeProps {
-  above?: Exclude<Breakpoint, 'desktop'>;
+  above?: Exclude<Breakpoint, 'wide'>;
   below?: Exclude<Breakpoint, 'mobile'>;
 }
 
 export const resolveResponsiveRangeProps = (
   props: ResponsiveRangeProps,
-): [boolean, boolean, boolean] => {
+): [boolean, boolean, boolean, boolean] => {
   const { above, below } = props;
 
   if (!above && !below) {
-    return [false, false, false];
+    return [false, false, false, false];
   }
 
   const startIndex = above ? breakpointNames.indexOf(above) + 1 : 0;
@@ -23,6 +23,7 @@ export const resolveResponsiveRangeProps = (
   const includeMobile = range.indexOf('mobile') >= 0;
   const includeTablet = range.indexOf('tablet') >= 0;
   const includeDesktop = range.indexOf('desktop') >= 0;
+  const includeWide = range.indexOf('wide') >= 0;
 
-  return [includeMobile, includeTablet, includeDesktop];
+  return [includeMobile, includeTablet, includeDesktop, includeWide];
 };